src/order/requirements.txt
paho-mqtt
prometheus_client
prometheus_flask_exporter 
Flask
pydantic
redis